* implemented Undo for both folder and file links

+++ CloudSyncHelper/Execution/FileLink.cs
@@ -30,7 +30,10 @@ namespace XElement.CloudSyncHelper.Execution
 
         public override void /*LinkBase.*/Undo()
         {
-            File.Delete( this.LinkPath );
+            if ( File.Exists( this.LinkPath ) )
+            {
+                File.Delete( this.LinkPath );
+            }
         }
     }
 #endregion
